requirements.txt
Flask==3.0.2
Jinja2==3.1.2
paho-mqtt==1.6.1
pymongo==4.11
requests==2.31.0
rpyc==6.0.1
